Tests - Move NWRC class to utils folder.
